+++ runtime/catalog-karapace/src/main/java/io/aklivity/zilla/runtime/catalog/karapace/internal/config/KarapaceOptionsConfigBuilder.java
@@ -12,7 +12,7 @@
  * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package io.aklivity.zilla.runtime.catalog.schema.registry.internal.config;
+package io.aklivity.zilla.runtime.catalog.karapace.internal.config;
 
 import java.time.Duration;
 import java.util.function.Function;
