+++ runtime/catalog-karapace/src/main/resources/META-INF/services/io.aklivity.zilla.runtime.engine.catalog.CatalogFactorySpi
@@ -1,1 +1,1 @@
-io.aklivity.zilla.runtime.catalog.schema.registry.internal.KarapaceCatalogFactorySpi
+io.aklivity.zilla.runtime.catalog.karapace.internal.KarapaceCatalogFactorySpi
