+++ runtime/catalog-karapace/src/main/resources/META-INF/services/io.aklivity.zilla.runtime.engine.config.OptionsConfigAdapterSpi
@@ -1,1 +1,1 @@
-io.aklivity.zilla.runtime.catalog.schema.registry.internal.config.KarapaceOptionsConfigAdapter
+io.aklivity.zilla.runtime.catalog.karapace.internal.config.KarapaceOptionsConfigAdapter
