+++ runtime/catalog-karapace/src/main/resources/META-INF/services/io.aklivity.zilla.runtime.engine.event.EventFormatterFactorySpi
@@ -1,1 +1,1 @@
-io.aklivity.zilla.runtime.catalog.schema.registry.internal.KarapaceEventFormatterFactory
+io.aklivity.zilla.runtime.catalog.karapace.internal.KarapaceEventFormatterFactory
